feat: add wavaeform visualizer

+++ src/util.rs
@@ -35,25 +35,25 @@ pub fn parse_bits(input: &str) -> Option<Box<[u8]>> {
 }
 
 #[inline]
-pub fn check_bit_period(tb: f32) -> anyhow::Result<f32> {
+pub fn check_bit_period(tb: f64) -> anyhow::Result<f64> {
     anyhow::ensure!(tb.is_finite() && tb > 0.0, CoderInitErr::InvalidBitPeriod);
     Ok(tb)
 }
 
 #[inline]
-pub fn check_ampl_closed(v: f32) -> anyhow::Result<f32> {
+pub fn check_ampl_closed(v: f64) -> anyhow::Result<f64> {
     anyhow::ensure!(v.is_finite() && v > 0.0, CoderInitErr::BadAmplitude);
     Ok(v)
 }
 
 #[inline]
-pub fn check_ampl_opened(v: f32) -> anyhow::Result<f32> {
+pub fn check_ampl_opened(v: f64) -> anyhow::Result<f64> {
     anyhow::ensure!(v.is_finite() && v != 0.0, CoderInitErr::BadAmplitude);
     Ok(v)
 }
 
 #[inline]
-pub fn check_duty(duty: f32) -> anyhow::Result<f32> {
+pub fn check_duty(duty: f64) -> anyhow::Result<f64> {
     anyhow::ensure!(
         duty.is_finite() && (duty > 0.0 || duty <= 1.0),
         CoderInitErr::WrongDuty
